Add updatecontent checks
Add checks to validate update content. Verify that the published files
match the hashes listed in the manifests and the manifest files match
the hashes listed in the MoM. Verify that the zero packs contain all
necessary full files for each bundle and the full files match the hashes
in the manifest.